Refactor: Move TabContent to a separate component
I moved the `TabContent` function from `MobileEditorModal.tsx` to its own file, `src/client/components/mobile/TabContent.tsx`.

This refactoring prevents the `TabContent` from being redefined on every render of `MobileEditorModal`, which was causing a `ReferenceError` on mobile devices.
